\name{epimisc-package}
\alias{epimisc-package}
\alias{epimisc}
\docType{package}
\title{
\packageTitle{epimisc}
}
\description{
\packageDescription{epimisc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{epimisc}
\packageIndices{epimisc}
}
\author{
\packageAuthor{epimisc}

Maintainer: \packageMaintainer{epimisc}
}
\keyword{ package }
